eval/requirements.txt
deepeval
google-genai
